Added admin bar icon, changed window resize to use debounce (instead of throttle)
New icon indicates when a page is deregistering assets so you don't need to hover over the edit menu to check. #2036

--- 

There is no guarantee that `throttle()` will trigger once at the very end, so detections may not be correct whereas `debounce()` does.

If we need to do both, we could, but I want to start with this for now.